Extracted buttons and typedefs
