+++ components/ReviewBox.tsx
@@ -1,3 +1,5 @@
+import PrimaryButton from './PrimaryButton'
+
 export default props => {
   let stars = ''
   for (let i = 0; i < props.stars; i++) {
@@ -16,13 +18,9 @@ export default props => {
 
         <span>{props.content}</span>
       </div>
-      <div className="w-full flex justify-between mb-4">
-        <button className="inline-block text-sm px-2 py-2 leading-none border rounded text-indigo-500 border-indigo-500 hover:border-transparent hover:text-white hover:bg-indigo-500 mt-4 ">
-          Not Helpful: {props.not_helpful_votes}
-        </button>
-        <button className="inline-block text-sm px-2 py-2 leading-none border rounded text-indigo-500 border-indigo-500 hover:border-transparent hover:text-white hover:bg-indigo-500 mt-4 ">
-          Helpful: {props.helpful_votes}
-        </button>
+      <div className="w-full flex justify-between">
+        <PrimaryButton buttonText={' Not Helpful: ' + props.not_helpful_votes} />
+        <PrimaryButton buttonText={' Helpful: ' + props.helpful_votes} />
       </div>
     </div>
   )
