+++ pages/api/index.ts
@@ -1,50 +1,8 @@
-import { ApolloServer, gql } from 'apollo-server-micro'
+import { ApolloServer } from 'apollo-server-micro'
 import Cors from 'micro-cors'
 import knex from 'knex'
 
-const typeDefs = gql`
-  type Query {
-    woods(first: Int = 25, skip: Int = 0): [Wood!]!
-    stains(first: Int = 25, skip: Int = 0): [Stain!]!
-    boards(first: Int = 25, skip: Int = 0): [Board!]!
-    board(id: ID!): Board!
-    reviews(first: Int = 25, skip: Int = 0): [Review!]!
-  }
-
-  type Wood {
-    id: ID!
-    name: String!
-  }
-
-  type Stain {
-    id: ID!
-    name: String!
-  }
-
-  type Board {
-    id: ID!
-    stain: Stain!
-    wood: Wood!
-    reviews(first: Int = 25, skip: Int = 0): [Review!]!
-    stock: Int!
-    width_in_cm: Int!
-    length_in_cm: Int!
-    thickness_in_cm: Int!
-    price_in_usd: Int!
-    picture_url: String!
-    description: String!
-  }
-
-  type Review {
-    id: ID!
-    user_id: Int!
-    board: Board!
-    stars: Int!
-    helpful_votes: Int!
-    not_helpful_votes: Int!
-    content: String!
-  }
-`
+import typeDefs from '../../constants/graphql/typeDefs'
 
 const resolvers = {
   Query: {
@@ -112,7 +70,6 @@ const resolvers = {
         .orderBy('helpful_votes', 'asc')
     }
   }
-
 }
 
 const db = knex({
